+++ Scripts/Items/Addons/TapestryOfSosaria.cs
@@ -66,7 +66,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.WriteEncodedInt((int)0); // version
+            writer.WriteEncodedInt(0); // version
 
             writer.WriteEncodedInt((int)this.m_Level);
         }
