+++ Scripts/Items/Artifacts/Equipment/Armor/SentinelsGuard.cs
@@ -82,7 +82,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)2);
+            writer.Write(2);
         }
 
         public override void Deserialize(GenericReader reader)
