+++ Scripts/Items/Artifacts/Equipment/Weapons/VampiricEssence.cs
@@ -40,7 +40,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0);
+            writer.Write(0);
         }
 
         public override void Deserialize(GenericReader reader)
