+++ Scripts/Items/Consumables/BaseWaterContainer.cs
@@ -118,8 +118,8 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
-            writer.Write((int)m_Quantity);
+            writer.Write(0); // version
+            writer.Write(m_Quantity);
         }
 
         public override void Deserialize(GenericReader reader)
